moving folders around - preparing to make schema name a first class citizen in url
