+++ UI-dev1.5/src/routes/serversServerDatabasesDatabaseTablesTable/index.js
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import TableRows from "./tableRows";
+import TableBrowse from "./subroutes/browse";
 import TableFields from "./subroutes/fields";
 import TableCreateSql from "./subroutes/createSql";
 import TableSql from "./subroutes/sql";
@@ -10,4 +10,4 @@ export default function ServersServerDatabasesDatabaseTablesTable(){
     return <Outlet />
 }
 
-export {TableRows,TableFields,TableCreateSql,TableSql};
+export {TableBrowse,TableFields,TableCreateSql,TableSql};
