+++ UI-dev1.5/src/routes/serversServerDatabasesDatabaseTablesTable/subroutes/browse/index.jsx
@@ -1,12 +1,12 @@
 import { useEffect } from "react";
 import { useDispatch,useSelector } from "react-redux";
-import useCurrents from "../../services/useCurrents";
-import { findConnectionNameByDbOrServer } from '../../store/dbTreeSlice';
-import { runQuery } from "../../store/querySlice";
-import {LastQuery,QueryEditor,QueryResults} from "../../components/query";
-import FieldsHelper from "./fieldsHelper";
+import useCurrents from "../../../../services/useCurrents";
+import { findConnectionNameByDbOrServer } from '../../../../store/dbTreeSlice';
+import { runQuery } from "../../../../store/querySlice";
+import {LastQuery,QueryEditor,QueryResults} from "../../../../components/query";
+import FieldsHelper from "../../fieldsHelper";
 
-export default function TableRows(){
+export default function TableBrowse(){
     const dispatch = useDispatch();
     const {server,database,table} = useCurrents();
     const connectionName = useSelector(findConnectionNameByDbOrServer(server,database));
